Reto2/Solucion/API_Gateway/requirements.txt
grpcio==1.43.0
grpcio-tools==1.43.0

